requirements.txt
Django==5.2.4

# 数据库支持
mysqlclient>=2.2.0  # MySQL数据库客户端

# API开发
djangorestframework>=3.14.0  # 用于API开发
django-cors-headers>=4.0.0  # 用于处理跨域请求

# 其他依赖可以在这里添加
# 例如：
# pillow>=10.0.0  # 用于图片处理